+++ vti/src/cli.ts
@@ -1,9 +1,13 @@
-import { Command } from 'commander';
-import { diagnostics } from './commands/diagnostics';
+import { Command, Option } from 'commander';
+import { diagnostics, LogLevel, logLevels } from './commands/diagnostics';
 
 function getVersion(): string {
   const { version }: { version: string } = require('../package.json');
   return `v${version}`;
+}
+
+function validateLogLevel(logLevelInput: unknown): logLevelInput is LogLevel {
+  return typeof logLevelInput === 'string' && (logLevels as ReadonlyArray<string>).includes(logLevelInput);
 }
 
 (async () => {
@@ -13,8 +17,18 @@ function getVersion(): string {
   program
     .command('diagnostics [workspace]')
     .description('Print all diagnostics')
-    .action(async workspace => {
-      await diagnostics(workspace);
+    .addOption(
+      new Option('-l, --log-level <logLevel>', 'Log level to print')
+        .default('WARN')
+        // logLevels is readonly array but .choices need read-write array (because of weak typing)
+        .choices((logLevels as unknown) as string[])
+    )
+    .action(async (workspace, options) => {
+      const logLevelOption: unknown = options.logLevel;
+      if (!validateLogLevel(logLevelOption)) {
+        throw new Error(`Invalid log level: ${logLevelOption}`);
+      }
+      await diagnostics(workspace, logLevelOption);
     });
 
   program.parse(process.argv);